refactor(core): rename admin support credits route to /credits
