+++ apps/core/src/routes/v1/admin/index.ts
@@ -2,10 +2,10 @@ import { createMiddleware } from "hono/factory";
 
 import { OpenAPIHonoWithAuth } from "@/lib/hono";
 import { requireAdminAuthContext } from "@/middleware/auth";
+import creditsRouter from "./credits/index.js";
 import invoicesRouter from "./invoices/index.js";
 import organizationsRouter from "./organizations/index.js";
 import searchRouter from "./search/index.js";
-import supportCreditsRouter from "./support-credits/index.js";
 import tasksRouter from "./tasks/index.js";
 import usersRouter from "./users/index.js";
 
@@ -21,7 +21,7 @@ app.route("/search", searchRouter);
 app.route("/users", usersRouter);
 app.route("/organizations", organizationsRouter);
 app.route("/invoices", invoicesRouter);
-app.route("/support-credits", supportCreditsRouter);
+app.route("/credits", creditsRouter);
 app.route("/tasks", tasksRouter);
 
 export default app;
